+++ src/rad/src/types.h
@@ -31,8 +31,8 @@ enum FaceFlags : unsigned {
 };
 
 struct Plane : public bsp::BSPPlane {
-    glm::vec3 vWorldOrigin;
-    glm::vec3 vI, vJ;
+    glm::dvec3 vWorldOrigin;
+    glm::dvec3 vI, vJ;
 
     //! List of faces that are in this plane
     std::vector<unsigned> faces;
@@ -54,9 +54,9 @@ struct Face : public bsp::BSPFace {
     unsigned iFlags = 0;
 
     glm::vec3 vNormal;                //! World normal vector. +/- pPlane->vNormal
-    glm::vec3 vWorldOrigin;           //!< Origin of plane/face coord system.
-    glm::vec3 vPlaneI, vPlaneJ;       //!< Plane X/Y in world coords.
-    glm::vec2 vFaceI, vFaceJ;         //!< Face axis in plane coords.
+    glm::dvec3 vWorldOrigin;           //!< Origin of plane/face coord system.
+    glm::dvec3 vPlaneI, vPlaneJ;       //!< Plane X/Y in world coords.
+    glm::dvec2 vFaceI, vFaceJ;         //!< Face axis in plane coords.
     glm::vec2 vPlaneMins, vPlaneMaxs; //!< Bounds in the plane coords.
     glm::vec2 vPlaneCenter;           //!< Center point in plane coords
     glm::vec2 vFaceMins, vFaceMaxs;   //!< Bounds in the face coords.
